Make path handling safer.
Previous versions of this code relied on the upload subdir configuration
parameter being set to a relative pathname, and then prepending a leading slash
to make it absolute. This is a little fragile, especially where the config
parameter is specified with a leading slash, so replace this with path.Join to
get consistent behaviour in both of these cases.

The store directory's pathname is also assumed to be given with a trailing
slash, with the request filename being simply appended to this; use
filepath.Join to get consistent behaviour here too, in case the trailing slash
is not provided.

Note that http.Server performs path canonicalisation and sanitisation on the
request URL path internally before invoking the specified handler function, so
this doesn't need to be done manually.