Add support for listening on Unix domain sockets.
This adds an optional boolean-valued configuration file option
"unixsocket". When specified, it changes the meaning of "listenport" to
indicate the path to a Unix domain socket on which prosody-filer should listen,
instead of a TCP port.

This is useful for multi-user shared hosting environments, where listening on
loopback TCP sockets may allow other, unrelated users to connect to the filer.